Add removeFill option for svg task with option mode 'use'
When I am a developer using SVG `use` syntax
And having control of the SVGs piped through the task
Then I should be able to choose if the task removes the fill's from the SVGs.

Currently if SVGs with a fill (such as flags) go through the pipeline, the fill is removed and ignores the `removeFill` option.